Add files via upload
These codes were used in my OJT project named LMMS App which stands for Lathe Machine Maintenance Sheet. This app is made as a substitute to traditional way of using paper check sheets for the Preventive Maintenance of Lathe Machines. I used Unity to develop the app and installed it on my phone. The application has four main functions/buttons: 

 Main Menu Scene has:
*LMMS button - contains the main check sheet.                                    (SaveInputFields)
*Summary Report button - displays the recent saved check sheet        (Summary)
*Tutorial Guides - contains three videos.                                                (GoBackHome)
* Notes - where you can save notepads.                                                (NotepadScripts)